+++ __tests__/utils/lngs-to-load-detector.test.js
@@ -1,36 +1,35 @@
 /* eslint-env jest */
+import lngsToLoadDetector from '../../src/utils/lngs-to-load-detector'
 
-import analyseFallbackLanguageConfiguration from '../../src/utils/fallback-lng-analysis'
-
-describe('analyseFallbackLanguageConfiguration utility function', () => {
+describe('lngsToLoadDetector utility function', () => {
 
   it('returns only initialLng if fallbackLng equals', () => {
-    const analysis = analyseFallbackLanguageConfiguration('en', 'en')
+    const analysis = lngsToLoadDetector('en', 'en')
     expect(analysis).toMatchObject(['en'])
   })
 
   it('returns both initialLng and fallbackLng if differs', () => {
-    const analysis = analyseFallbackLanguageConfiguration('de', 'en')
+    const analysis = lngsToLoadDetector('en', 'de')
     expect(analysis).toMatchObject(['en', 'de'])
   })
 
   it('returns locale specific fallbackLng', () => {
-    const analysis = analyseFallbackLanguageConfiguration({ 'en-BE': 'en' }, 'en-BE')
+    const analysis = lngsToLoadDetector('en-BE', { 'en-BE': 'en' })
     expect(analysis).toMatchObject(['en-BE', 'en'])
   })
 
   it('returns locale specific fallbackLng and default', () => {
-    const analysis = analyseFallbackLanguageConfiguration({ 'de-BE': 'de', default: 'en' }, 'de-BE')
+    const analysis = lngsToLoadDetector('de-BE', { 'de-BE': 'de', default: 'en' })
     expect(analysis).toMatchObject(['de-BE', 'de', 'en'])
   })
 
   it('returns locale specific fallbackLng array', () => {
-    const analysis = analyseFallbackLanguageConfiguration({ 'fi-FI': ['fi', 'en'] }, 'fi-FI')
+    const analysis = lngsToLoadDetector('fi-FI', { 'fi-FI': ['fi', 'en'] })
     expect(analysis).toMatchObject(['fi-FI', 'fi', 'en'])
   })
 
   it('returns locale specific fallbackLng array and default', () => {
-    const analysis = analyseFallbackLanguageConfiguration({ 'fi-FI': ['fi', 'en'], default: 'it' }, 'fi-FI')
+    const analysis = lngsToLoadDetector('fi-FI', { 'fi-FI': ['fi', 'en'], default: 'it' })
     expect(analysis).toMatchObject(['fi-FI', 'fi', 'en', 'it'])
   })
